Coercing nil values to false for boolean fields
This change adds some code in the `:stillsuit/ref` resolver to ensure
that fields declared as Boolean (in the resolver args) always return
either true or false (and not nil).

I also added a little convenience wrapper function to execute a query
given the results of `(stillsuit/decorate)`, and rearranged a bit of
the test code slightly around that.